+++ test/Parity Tests/Aggregate Tests/Aggregate Tests.js
@@ -39,6 +39,7 @@ module.exports = function ( Driver )
 		require( './test-suite/Accumulator Operator Tests.js' )( Driver );
 		require( './test-suite/Reshaping Stage Tests.js' )( Driver );
 		require( './test-suite/Bucketing Stage Tests.js' )( Driver );
+		require( './test-suite/Filling Stage Tests.js' )( Driver );
 		require( './test-suite/Stage and Accumulator Tests.js' )( Driver );
 		require( './test-suite/Expression Rejection Tests.js' )( Driver );
 	} );